CommandBase duplicate removed, pretty dict printing moved to the only one
